+++ tests/integration/x/vm/test_call_evm.go
@@ -10,25 +10,39 @@ import (
 	testconstants "github.com/cosmos/evm/testutil/constants"
 	utiltx "github.com/cosmos/evm/testutil/tx"
 	"github.com/cosmos/evm/x/erc20/types"
+	"github.com/cosmos/evm/x/vm/statedb"
 	evmtypes "github.com/cosmos/evm/x/vm/types"
 )
 
 func (s *KeeperTestSuite) TestCallEVM() {
 	wcosmosEVMContract := common.HexToAddress(testconstants.WEVMOSContractMainnet)
 	testCases := []struct {
-		name    string
-		method  string
-		expPass bool
+		name     string
+		method   string
+		stateDB  *statedb.StateDB
+		expPass  bool
+		expError string
 	}{
 		{
 			"unknown method",
 			"",
+			nil,
 			false,
+			"",
 		},
 		{
 			"pass",
 			"balanceOf",
+			nil,
 			true,
+			"",
+		},
+		{
+			"fail with nil statedb",
+			"balanceOf",
+			nil,
+			false,
+			"stateDB cannot be nil",
 		},
 	}
 	for _, tc := range testCases {
@@ -36,12 +50,23 @@ func (s *KeeperTestSuite) TestCallEVM() {
 
 		erc20 := contracts.ERC20MinterBurnerDecimalsContract.ABI
 		account := utiltx.GenerateAddress()
-		res, err := s.Network.App.GetEVMKeeper().CallEVM(s.Network.GetContext(), erc20, types.ModuleAddress, wcosmosEVMContract, false, nil, tc.method, account)
+
+		var stateDB *statedb.StateDB
+		if tc.stateDB == nil && tc.name != "fail with nil statedb" {
+			stateDB = statedb.New(s.Network.GetContext(), s.Network.App.GetEVMKeeper(), statedb.NewEmptyTxConfig())
+		} else {
+			stateDB = tc.stateDB
+		}
+
+		res, err := s.Network.App.GetEVMKeeper().CallEVM(s.Network.GetContext(), stateDB, erc20, types.ModuleAddress, wcosmosEVMContract, false, false, nil, tc.method, account)
 		if tc.expPass {
 			s.Require().IsTypef(&evmtypes.MsgEthereumTxResponse{}, res, tc.name)
 			s.Require().NoError(err)
 		} else {
 			s.Require().Error(err)
+			if tc.expError != "" {
+				s.Require().Contains(err.Error(), tc.expError)
+			}
 		}
 	}
 }
@@ -54,82 +79,88 @@ func (s *KeeperTestSuite) TestCallEVMWithData() {
 		from     common.Address
 		malleate func() []byte
 		deploy   bool
-		gasCap   *big.Int
+		useNilDB bool
 		expPass  bool
+		expError string
 	}{
 		{
-			"pass with unknown method",
-			types.ModuleAddress,
-			func() []byte {
+			name: "pass with unknown method",
+			from: types.ModuleAddress,
+			malleate: func() []byte {
 				account := utiltx.GenerateAddress()
 				data, _ := erc20.Pack("", account)
 				return data
 			},
-			false,
-			nil,
-			true,
+			deploy:   false,
+			useNilDB: false,
+			expPass:  true,
+			expError: "",
 		},
 		{
-			"pass",
-			types.ModuleAddress,
-			func() []byte {
+			name: "pass",
+			from: types.ModuleAddress,
+			malleate: func() []byte {
 				account := utiltx.GenerateAddress()
 				data, _ := erc20.Pack("balanceOf", account)
 				return data
 			},
-			false,
-			nil,
-			true,
+			deploy:   false,
+			useNilDB: false,
+			expPass:  true,
+			expError: "",
 		},
 		{
-			"fail with small gas cap",
-			types.ModuleAddress,
-			func() []byte {
+			name: "pass with empty data",
+			from: types.ModuleAddress,
+			malleate: func() []byte {
+				return []byte{}
+			},
+			deploy:   false,
+			useNilDB: false,
+			expPass:  true,
+			expError: "",
+		},
+		{
+			name: "fail empty sender",
+			from: common.Address{},
+			malleate: func() []byte {
+				return []byte{}
+			},
+			deploy:   false,
+			useNilDB: false,
+			expPass:  false,
+			expError: "",
+		},
+		{
+			name: "fail with nil statedb",
+			from: types.ModuleAddress,
+			malleate: func() []byte {
 				account := utiltx.GenerateAddress()
 				data, _ := erc20.Pack("balanceOf", account)
 				return data
 			},
-			false,
-			big.NewInt(1),
-			false,
+			deploy:   false,
+			useNilDB: true,
+			expPass:  false,
+			expError: "stateDB cannot be nil",
 		},
 		{
-			"pass with empty data",
-			types.ModuleAddress,
-			func() []byte {
-				return []byte{}
-			},
-			false,
-			nil,
-			true,
-		},
-
-		{
-			"fail empty sender",
-			common.Address{},
-			func() []byte {
-				return []byte{}
-			},
-			false,
-			nil,
-			false,
-		},
-		{
-			"deploy",
-			types.ModuleAddress,
-			func() []byte {
+			name: "deploy",
+			from: types.ModuleAddress,
+			malleate: func() []byte {
 				ctorArgs, _ := contracts.ERC20MinterBurnerDecimalsContract.ABI.Pack("", "test", "test", uint8(18))
 				data := append(contracts.ERC20MinterBurnerDecimalsContract.Bin, ctorArgs...) //nolint:gocritic
 				return data
 			},
-			true,
-			nil,
-			true,
+			deploy:   true,
+			useNilDB: false,
+			expPass:  true,
+			expError: "",
 		},
 		{
-			"fail deploy",
-			types.ModuleAddress,
-			func() []byte {
+			name: "fail deploy",
+			from: types.ModuleAddress,
+			malleate: func() []byte {
 				params := s.Network.App.GetEVMKeeper().GetParams(s.Network.GetContext())
 				params.AccessControl.Create = evmtypes.AccessControlType{
 					AccessType: evmtypes.AccessTypeRestricted,
@@ -139,9 +170,23 @@ func (s *KeeperTestSuite) TestCallEVMWithData() {
 				data := append(contracts.ERC20MinterBurnerDecimalsContract.Bin, ctorArgs...) //nolint:gocritic
 				return data
 			},
-			true,
-			nil,
-			false,
+			deploy:   true,
+			useNilDB: false,
+			expPass:  false,
+			expError: "",
+		},
+		{
+			name: "fail deploy with nil statedb",
+			from: types.ModuleAddress,
+			malleate: func() []byte {
+				ctorArgs, _ := contracts.ERC20MinterBurnerDecimalsContract.ABI.Pack("", "test", "test", uint8(18))
+				data := append(contracts.ERC20MinterBurnerDecimalsContract.Bin, ctorArgs...) //nolint:gocritic
+				return data
+			},
+			deploy:   true,
+			useNilDB: true,
+			expPass:  false,
+			expError: "stateDB cannot be nil",
 		},
 	}
 
@@ -153,10 +198,15 @@ func (s *KeeperTestSuite) TestCallEVMWithData() {
 			var res *evmtypes.MsgEthereumTxResponse
 			var err error
 
+			var stateDB *statedb.StateDB
+			if !tc.useNilDB {
+				stateDB = statedb.New(s.Network.GetContext(), s.Network.App.GetEVMKeeper(), statedb.NewEmptyTxConfig())
+			}
+
 			if tc.deploy {
-				res, err = s.Network.App.GetEVMKeeper().CallEVMWithData(s.Network.GetContext(), tc.from, nil, data, true, tc.gasCap)
+				res, err = s.Network.App.GetEVMKeeper().CallEVMWithData(s.Network.GetContext(), stateDB, tc.from, nil, data, true, false, nil)
 			} else {
-				res, err = s.Network.App.GetEVMKeeper().CallEVMWithData(s.Network.GetContext(), tc.from, &wcosmosEVMContract, data, false, tc.gasCap)
+				res, err = s.Network.App.GetEVMKeeper().CallEVMWithData(s.Network.GetContext(), stateDB, tc.from, &wcosmosEVMContract, data, false, false, nil)
 			}
 
 			if tc.expPass {
@@ -164,6 +214,9 @@ func (s *KeeperTestSuite) TestCallEVMWithData() {
 				s.Require().NoError(err)
 			} else {
 				s.Require().Error(err)
+				if tc.expError != "" {
+					s.Require().Contains(err.Error(), tc.expError)
+				}
 			}
 		})
 	}