+++ CHANGELOG.md
@@ -45,6 +45,23 @@
 - [\#992](https://github.com/cosmos/evm/pull/992) Respect the provided `gasCap` in `CallEVMWithData` instead of always used the default cap.
 - [\#993](https://github.com/cosmos/evm/pull/993) Enforce `src_callback` contract address to match the packet sender for IBC acknowledgement and timeout callbacks to prevent arbitrary contract execution.
 
+## v0.6.0
+
+### DEPENDENCIES
+
+### BUG FIXES
+
+- Add `stateDB` and `callFromPrecompile` parameters to internal EVM messages.
+- Fixed an issue with events missing from final transaction result.
+
+### IMPROVEMENTS
+
+### FEATURES
+
+### STATE BREAKING
+
+### API-BREAKING
+
 ## v0.5.0
 
 ### DEPENDENCIES